update dependency: nx >= 2.0
recent versions of networkx changed the API regarding parameter order of
several functions (e.g.~set_node_attributes) in a way that is not
backwards compatible, in particular with the version from python-networkx
packages on current debian testing.

updating the dependency to v >=2 should cause setuptools to fetch a more
recent version of nx if necessary.